+++ test/Unit/FieldDefinitionTest.php
@@ -36,7 +36,7 @@ final class FieldDefinitionTest extends AbstractTestCase
     {
         $closure = static function (FixtureFactory $fixtureFactory): Fixture\FixtureFactory\Entity\User {
             /** @var Fixture\FixtureFactory\Entity\User $user */
-            $user = $fixtureFactory->get(Fixture\FixtureFactory\Entity\User::class);
+            $user = $fixtureFactory->create(Fixture\FixtureFactory\Entity\User::class);
 
             $user->renameTo(self::faker()->userName);