+++ CHANGELOG.md
@@ -36,6 +36,7 @@ For a full diff see [`fa9c564...master`][fa9c564...master].
 * Extracted `FieldDefinition\Closure` ([#161]), by [@localheinz]
 * Extracted `FieldDefinition\Sequence` ([#164]), by [@localheinz]
 * Introduced named constructors for field definitions and marked primary constructor as `private` ([#188]), by [@localheinz]
+* Renamed `FixtureFactory::get()` to `FixtureFactory::create()` ([#189]), by [@localheinz]
 
 ### Fixed
 
@@ -95,5 +96,6 @@ For a full diff see [`fa9c564...master`][fa9c564...master].
 [#164]: https://github.com/ergebnis/factory-bot/pull/164
 [#185]: https://github.com/ergebnis/factory-bot/pull/185
 [#188]: https://github.com/ergebnis/factory-bot/pull/188
+[#189]: https://github.com/ergebnis/factory-bot/pull/189
 
 [@localheinz]: https://github.com/localheinz